app working, finally.
